+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.1] 2023-11-09
+### Added
+- Unittests for hf model.
+
+### Fixed
+- Fix sample path for PyPI package.
+
 ## [0.1.0] 2023-11-09
 ### Added
 - First version of the conette package to load CoNeTTE model.